1.1
Rewritten iOS audio input/output

The SuperpoweredIOSAudioOutput class is now SuperpoweredIOSAudioIO.
Audio input can be set with the appropriate audio session category.
Routing changes are handled with less delay, and asking for recording
permissions are built-in.

Android improvements

Stream type can be set for the SuperpoweredAndroidAudioIO, such as
SL_ANDROID_STREAM_MEDIA for normal playback and SL_ANDROID_STREAM_VOICE
for the speakerphone.